Move generate_password to user_authn.
