+++ openedx/core/djangoapps/user_api/accounts/forms.py
@@ -7,7 +7,7 @@ from __future__ import absolute_import
 from django import forms
 from django.core.exceptions import ValidationError
 
-from openedx.core.djangoapps.user_api.accounts.utils import generate_password
+from openedx.core.djangoapps.user_authn.utils import generate_password
 
 
 class RetirementQueueDeletionForm(forms.Form):
