+++ openedx/core/djangoapps/user_api/accounts/utils.py
@@ -3,9 +3,7 @@ Utility methods for the account settings.
 """
 from __future__ import absolute_import, unicode_literals
 
-import random
 import re
-import string
 
 import waffle
 from completion import waffle as completion_waffle
@@ -182,20 +180,6 @@ def retrieve_last_sitewide_block_completed(user):
         course_key=text_type(item.location.course_key),
         location=text_type(item.location),
     )
-
-
-def generate_password(length=12, chars=string.ascii_letters + string.digits):
-    """Generate a valid random password"""
-    if length < 8:
-        raise ValueError("password must be at least 8 characters")
-
-    choice = random.SystemRandom().choice
-
-    password = ''
-    password += choice(string.digits)
-    password += choice(string.ascii_letters)
-    password += ''.join([choice(chars) for _i in range(length - 2)])
-    return password
 
 
 def is_secondary_email_feature_enabled():
