+++ openedx/core/djangoapps/user_api/management/commands/cancel_user_retirement_request.py
@@ -9,8 +9,8 @@ import logging
 
 from django.core.management.base import BaseCommand, CommandError
 
-from openedx.core.djangoapps.user_api.accounts.utils import generate_password
 from openedx.core.djangoapps.user_api.models import UserRetirementStatus
+from openedx.core.djangoapps.user_authn.utils import generate_password
 
 LOGGER = logging.getLogger(__name__)
