+++ openedx/core/djangoapps/user_authn/tests/test_utils.py
@@ -11,7 +11,9 @@ from django.test.utils import override_settings
 from six.moves.urllib.parse import urlencode  # pylint: disable=import-error
 
 from openedx.core.djangoapps.oauth_dispatch.tests.factories import ApplicationFactory
-from openedx.core.djangoapps.user_authn.utils import is_safe_login_or_logout_redirect
+from openedx.core.djangoapps.user_authn.utils import (
+    generate_password, is_safe_login_or_logout_redirect
+)
 
 
 @ddt.ddt
@@ -61,3 +63,29 @@ class TestRedirectUtils(TestCase):
         req = self.request.get('/logout?{}'.format(urlencode(params)), HTTP_HOST=host)
         actual_is_safe = is_safe_login_or_logout_redirect(req, redirect_url)
         self.assertEqual(actual_is_safe, expected_is_safe)
+
+
+class GeneratePasswordTest(TestCase):
+    """Tests formation of randomly generated passwords."""
+
+    def test_default_args(self):
+        password = generate_password()
+        self.assertEqual(12, len(password))
+        self.assertTrue(any(c.isdigit for c in password))
+        self.assertTrue(any(c.isalpha for c in password))
+
+    def test_length(self):
+        length = 25
+        self.assertEqual(length, len(generate_password(length=length)))
+
+    def test_chars(self):
+        char = '!'
+        password = generate_password(length=12, chars=(char,))
+
+        self.assertTrue(any(c.isdigit for c in password))
+        self.assertTrue(any(c.isalpha for c in password))
+        self.assertEqual(char * 10, password[2:])
+
+    def test_min_length(self):
+        with self.assertRaises(ValueError):
+            generate_password(length=7)
