+++ openedx/core/djangoapps/user_authn/utils.py
@@ -3,6 +3,9 @@ Utility functions used during user authentication.
 """
 from __future__ import absolute_import
 
+import random
+import string
+
 from django.conf import settings
 from django.utils import http
 from oauth2_provider.models import Application
@@ -29,3 +32,17 @@ def is_safe_login_or_logout_redirect(request, redirect_to):
         redirect_to, allowed_hosts=login_redirect_whitelist, require_https=request.is_secure(),
     )
     return is_safe_url
+
+
+def generate_password(length=12, chars=string.ascii_letters + string.digits):
+    """Generate a valid random password"""
+    if length < 8:
+        raise ValueError("password must be at least 8 characters")
+
+    choice = random.SystemRandom().choice
+
+    password = ''
+    password += choice(string.digits)
+    password += choice(string.ascii_letters)
+    password += ''.join([choice(chars) for _i in range(length - 2)])
+    return password
